backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pypdf==3.17.4
python-multipart==0.0.6
Pillow>=10.0.0
sqlmodel==0.0.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
docx2pdf==0.1.8

